requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.8.0
arclet-alconna==1.8.11
arclet-alconna-tools==0.7.3
asgiref==3.8.1
astor==0.8.1
attrs==25.1.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.3
bili_ticket_gt_python==0.2.7
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.5.5
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
decorator==5.2.1
Deprecated==1.2.18
distro==1.9.0
dnspython==2.6.1
duckdb==1.2.0
durationpy==0.9
email_validator==2.1.1
fastapi==0.115.12
fastapi-cli==0.0.3
ffmpy==0.5.0
filelock==3.17.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
google==3.0.0
google-auth==2.38.0
googleapis-common-protos==1.68.0
gradio==5.23.3
gradio_client==1.8.0
graphviz==0.20.3
greenlet==3.0.3
groovy==0.1.2
grpcio==1.70.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.29.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
Jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kdl==0.2.21
kiwisolver==1.4.8
kubernetes==32.0.1
langchain==0.3.19
langchain-core==0.3.40
langchain-openai==0.3.7
langchain-text-splitters==0.3.6
langsmith==0.3.11
loguru==0.7.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.7.5
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
narwhals==1.33.0
nepattern==0.7.2
nonebot-adapter-console==0.5.0
nonebot-adapter-onebot==2.4.3
nonebot-plugin-alconna==0.45.4
nonebot-plugin-htmlrender==0.3.1
nonebot2==2.3.0
nonechat==0.2.1
noneprompt==0.1.9
ntplib==0.4.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.65.1
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
orjson==3.10.3
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandasai==3.0.0b14
pandasai-openai==0.1.5
pefile==2023.2.7
pillow==10.4.0
playwright==1.43.0
plotly==6.0.1
posthog==3.18.0
propcache==0.3.0
protobuf==5.29.3
py==1.11.0
py-machineid==0.6.0
pyarrow==14.0.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.9.9
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pydotplus==2.0.2
pydub==0.25.1
pyee==11.1.0
Pygments==2.19.1
pygtrie==2.5.0
pyinstaller==6.4.0
pyinstaller-hooks-contrib==2024.1
pymdown-extensions==10.8.1
PyMySQL==1.1.1
pynvml==11.5.0
pyparsing==3.2.1
pyperclip==1.9.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pystun3==2.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-markdown-math==0.8
python-multipart==0.0.20
pytz==2020.5
pywin32-ctypes==0.2.2
PyYAML==6.0.2
qrcode==8.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retry==0.9.2
rich==13.7.1
rpds-py==0.24.0
rsa==4.9
ruff==0.11.4
safehttpx==0.1.6
scipy==1.10.1
seaborn==0.12.2
semantic-version==2.10.0
sentry-sdk==2.18.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
sqlglot==25.34.1
starlette==0.46.1
streamlit==1.44.1
sympy==1.13.3
tarina==0.4.4
tenacity==9.0.0
textual==0.29.0
tiktoken==0.9.0
tokenizers==0.21.0
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.67.1
typer==0.12.3
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.29.0
volcengine==1.0.174
watchdog==6.0.0
watchfiles==0.21.0
websocket-client==1.8.0
websockets==12.0
win32-setctime==1.1.0
winregistry==1.1.1
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
